healthcare_eval_framework_free/requirements.txt
trustllm>=0.3.0
trulens-eval>=1.4.0
transformers>=4.40.0
datasets>=2.18.0
evaluate>=0.4.1
sentencepiece
torch
psutil
tqdm
pyyaml
typer[all]
